Issue#95 - Provide update of server group targeting
