+++ core/src/main/python/wlsdeploy/tool/create/domain_creator.py
@@ -32,7 +32,6 @@ from wlsdeploy.aliases.model_constants import RESOURCE_GROUP_TEMPLATE
 from wlsdeploy.aliases.model_constants import SECURITY
 from wlsdeploy.aliases.model_constants import SECURITY_CONFIGURATION
 from wlsdeploy.aliases.model_constants import SERVER
-from wlsdeploy.aliases.model_constants import SERVER_GROUP_TARGETING_LIMITS
 from wlsdeploy.aliases.model_constants import SERVER_START_MODE
 from wlsdeploy.aliases.model_constants import SERVER_TEMPLATE
 from wlsdeploy.aliases.model_constants import SET_OPTION_APP_DIR
@@ -50,6 +49,7 @@ from wlsdeploy.tool.create.security_provider_creator import SecurityProviderCrea
 from wlsdeploy.tool.deploy import model_deployer
 from wlsdeploy.tool.util.archive_helper import ArchiveHelper
 from wlsdeploy.tool.util.library_helper import LibraryHelper
+from wlsdeploy.tool.util.target_helper import TargetHelper
 from wlsdeploy.tool.util.topology_helper import TopologyHelper
 from wlsdeploy.util import dictionary_utils
 from wlsdeploy.util import model as model_helper
@@ -105,6 +105,9 @@ class DomainCreator(Creator):
 
         self.library_helper = LibraryHelper(self.model, self.model_context, self.aliases, self._domain_home,
                                             ExceptionType.CREATE, self.logger)
+
+        self.target_helper = TargetHelper(self.model, self.model_context, self.aliases, ExceptionType.CREATE,
+                                          self.logger)
 
         #
         # Creating domains with the wls.jar template is busted for pre-12.1.2 domains with regards to the
@@ -303,7 +306,7 @@ class DomainCreator(Creator):
         self.__configure_fmw_infra_database()
 
         server_groups_to_target = self._domain_typedef.get_server_groups_to_target()
-        self.__target_server_groups_to_servers(server_groups_to_target)
+        self.target_helper.target_server_groups_to_servers(server_groups_to_target)
 
         self.logger.info('WLSDPLY-12209', self._domain_name,
                          class_name=self.__class_name, method_name=_method_name)
@@ -345,7 +348,7 @@ class DomainCreator(Creator):
             self.__configure_fmw_infra_database()
 
             server_groups_to_target = self._domain_typedef.get_server_groups_to_target()
-            self.__target_server_groups_to_servers(server_groups_to_target)
+            self.target_helper.target_server_groups_to_servers(server_groups_to_target)
 
         self.logger.info('WLSDPLY-12206', self._domain_name, domain_home,
                          class_name=self.__class_name, method_name=_method_name)
@@ -607,67 +610,6 @@ class DomainCreator(Creator):
 
         self.logger.info('WLSDPLY-12223', class_name=self.__class_name, method_name=_method_name)
         self.wlst_helper.get_database_defaults()
-
-        self.logger.exiting(class_name=self.__class_name, method_name=_method_name)
-        return
-
-    def __target_server_groups_to_servers(self, server_groups_to_target):
-        """
-        Target the server groups to the servers.
-        :param server_groups_to_target: the list of server groups to target
-        :raises: CreateException: if an error occurs
-        """
-        _method_name = '__target_server_groups_to_servers'
-
-        self.logger.entering(server_groups_to_target, class_name=self.__class_name, method_name=_method_name)
-        if len(server_groups_to_target) == 0:
-            return
-
-        location = LocationContext()
-        root_path = self.alias_helper.get_wlst_attributes_path(location)
-        self.wlst_helper.cd(root_path)
-
-        # We need to get the effective list of servers for the domain.  Since any servers
-        # referenced in the model have already been created but the templates may have
-        # defined new servers not listed in the model, get the list from WLST.
-        server_names = self._get_existing_server_names()
-
-        # Get the clusters and and their members
-        cluster_map = self._get_clusters_and_members_map()
-
-        # Get any limits that may have been defined in the model
-        domain_info = self.model.get_model_domain_info()
-        server_group_targeting_limits = \
-            dictionary_utils.get_dictionary_element(domain_info, SERVER_GROUP_TARGETING_LIMITS)
-        if len(server_group_targeting_limits) > 0:
-            server_group_targeting_limits = \
-                self._get_server_group_targeting_limits(server_group_targeting_limits, cluster_map)
-
-        # Get the map of server names to server groups to target
-        server_to_server_groups_map = self._get_server_to_server_groups_map(self._admin_server_name,
-                                                                            server_names,
-                                                                            server_groups_to_target,
-                                                                            server_group_targeting_limits)
-        if len(server_names) > 1:
-            for server, server_groups in server_to_server_groups_map.iteritems():
-                if len(server_groups) > 0:
-                    server_name = self.wlst_helper.get_quoted_name_for_wlst(server)
-                    self.logger.info('WLSDPLY-12224', str(server_groups), server_name,
-                                     class_name=self.__class_name, method_name=_method_name)
-                    self.wlst_helper.set_server_groups(server_name, server_groups)
-
-        elif len(server_group_targeting_limits) == 0:
-            #
-            # Domain has no managed servers and there were not targeting limits specified to target
-            # server groups to the admin server so make sure that the server groups are targeted to
-            # the admin server.
-            #
-            # This is really a best effort attempt.  It works for JRF domains but it is certainly possible
-            # that it may cause problems with other custom domain types.  Of course, creating a domain with
-            # no managed servers is not a primary use case of this tool so do it and hope for the best...
-            #
-            server_name = self.wlst_helper.get_quoted_name_for_wlst(server_names[0])
-            self.wlst_helper.set_server_groups(server_name, server_groups_to_target)
 
         self.logger.exiting(class_name=self.__class_name, method_name=_method_name)
         return
